테이블뷰 리스트 추가제거 버전 입니다..
23page

+++ Ex1.2.1_tableView_catalog/CartItem.h
@@ -6,12 +6,14 @@
 //  Copyright (c) 2014년 SDT-1. All rights reserved.
 //
 
+
 #import <Foundation/Foundation.h>
 #import "Product.h"
 
 @interface CartItem : NSObject
 
-@property (weak,nonatomic)Product *product;
-@property (nonatomic)NSInteger quantity;
+@property (weak, nonatomic)Product *product;
+@property (nonatomic) int quantity;
+
 
 @end
